worker/pyproject.toml
[project]
name = "babel-worker"
version = "0.1.0"
description = "Pull-worker for babel — runs translations locally against llama-server and streams results to the babel backend"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
]

[project.optional-dependencies]
tray = [
    "pystray>=0.19",
    "pillow>=10",
]

[project.scripts]
babel-worker = "babel_worker.cli:main"

[tool.setuptools]
packages = ["babel_worker"]
